Исправления линтерами. Изменения конфигурации .toml (exclude). Переименование папки test в tests
