+++ pyproject.toml
@@ -29,6 +29,7 @@ build-backend = "poetry.core.masonry.api"
 disallow_untyped_defs = true
 no_implicit_optional = true
 warn_return_any = true
+exclude = 'test_*'
 
 [tool.black]
 line-length = 119
