+++ tests/test_processing.py
@@ -1,0 +1,54 @@
+import pytest
+
+from src.processing import get_by_date_operations, get_state
+
+
+@pytest.mark.parametrize(
+    "state, expected",
+    [
+        (
+            "EXECUTED",
+            [
+                {"id": 41428829, "state": "EXECUTED", "date": "2019-07-03T18:35:29.512364"},
+                {"id": 939719570, "state": "EXECUTED", "date": "2018-06-30T02:08:58.425572"},
+            ],
+        ),
+        (
+            "CANCELED",
+            [
+                {"id": 594226727, "state": "CANCELED", "date": "2018-09-12T21:27:25.241689"},
+                {"id": 615064591, "state": "CANCELED", "date": "2018-10-14T08:21:33.419441"},
+            ],
+        ),
+        ("NOT IN STATE", []),
+    ],
+)
+def test_get_state(operations, state, expected):
+    assert get_state(operations, state) == expected
+
+
+@pytest.mark.parametrize(
+    "new_first, expected",
+    [
+        (
+            True,
+            [
+                {"id": 41428829, "state": "EXECUTED", "date": "2019-07-03T18:35:29.512364"},
+                {"id": 615064591, "state": "CANCELED", "date": "2018-10-14T08:21:33.419441"},
+                {"id": 594226727, "state": "CANCELED", "date": "2018-09-12T21:27:25.241689"},
+                {"id": 939719570, "state": "EXECUTED", "date": "2018-06-30T02:08:58.425572"},
+            ],
+        ),
+        (
+            False,
+            [
+                {"id": 939719570, "state": "EXECUTED", "date": "2018-06-30T02:08:58.425572"},
+                {"id": 594226727, "state": "CANCELED", "date": "2018-09-12T21:27:25.241689"},
+                {"id": 615064591, "state": "CANCELED", "date": "2018-10-14T08:21:33.419441"},
+                {"id": 41428829, "state": "EXECUTED", "date": "2019-07-03T18:35:29.512364"},
+            ],
+        ),
+    ],
+)
+def test_get_by_date_operations(operations, new_first, expected):
+    assert get_by_date_operations(operations, new_first) == expected
